feat: remove BroadcastStatusToPeersNode (step 3) and delete peer_broadcast_bt infrastructure
Remove the pre-SYNC raw Add(ParticipantStatus) re-broadcast step from
add_participant_status_tree and delete the now-unused broadcast infrastructure.

The Announce(CaseLedgerEntry) fan-out in _commit_log_cascade_bt already
propagates status updates to all participants (including the original
sender), making step 3's raw re-broadcast a redundant pre-SYNC artefact
that causes every peer to process the same update twice.

Changes:
- Remove BroadcastStatusToPeersNode from step 3 of add_participant_status_tree
- Simplify status/nodes/broadcast.py to only _find_case_manager_id (still
  needed by lifecycle.py for auto-close fan-out); define inline rather
  than re-exporting from the deleted broadcast package
- Remove broadcast node exports from status/nodes/__init__.py
- Delete vultron/core/behaviors/broadcast/ package (peer_broadcast_bt,
  generic broadcast nodes) — no remaining callers after step 3 removal
- Delete test/core/behaviors/broadcast/ and test_broadcast.py
- Remove TestBroadcastStatusToPeersNode and related sections from
  test_add_participant_status_bt.py; update one test assertion to match
  PR#1046 TerminateEmbargoNode fail-fast semantics

Per specs/multi-actor-demo.yaml DEMOMA-07-005. Closes #1045.
