+++ test/core/behaviors/status/test_add_participant_status_bt.py
@@ -15,19 +15,15 @@
 
 """Tests for AddParticipantStatus BT nodes and tree factory.
 
-Covers all five DEMOMA-07-003 steps:
+Covers all four DEMOMA-07-003 steps (step 3 raw re-broadcast removed
+per DEMOMA-07-005):
   1. VerifySenderIsParticipantNode — unknown sender is rejected
   2. AppendParticipantStatusNode  — status appended, RM regression rejected
-  3. BroadcastStatusToPeersNode   — fail-fast (BT-14-001): FAILURE when no
-     factory or delivery fails; SUCCESS when no peers to notify (no-op)
-  4. PublicDisclosureBranchNode   — SUCCESS when skip conditions met; FAILURE
-     when teardown needed but dispatch fails (BT-14-001)
+  4. PublicDisclosureBranchNode   — always SUCCESS, only triggers teardown on CS.P + CASE_OWNER
   5. AutoCloseBranchNode          — always SUCCESS, logs when all RM.CLOSED
 
-Per specs/multi-actor-demo.yaml DEMOMA-07-003.
+Per specs/multi-actor-demo.yaml DEMOMA-07-003 and DEMOMA-07-005.
 """
-
-from unittest.mock import MagicMock
 
 import py_trees
 import pytest
@@ -41,15 +37,10 @@ from vultron.core.behaviors.status.add_participant_status_tree import (
 from vultron.core.behaviors.status.append_participant_status_tree import (
     append_participant_status_tree,
 )
-from vultron.core.behaviors.broadcast.nodes import CreateBroadcastActivityNode
 from vultron.core.behaviors.status.nodes import (
     AppendStatusAndSaveParticipantNode,
     AutoCloseBranchNode,
-    BroadcastQueueToOutboxNode,
-    BroadcastStatusToPeersNode,
     CheckStatusNotAlreadyAppendedNode,
-    FilterPeerRecipientsNode,
-    FindCaseManagerNode,
     LoadParticipantNode,
     PublicDisclosureBranchNode,
     ResolveAndPersistStatusObjectNode,
@@ -644,360 +635,6 @@ class TestAppendParticipantStatusSubtree:
 
 
 # ---------------------------------------------------------------------------
-# Step 3: BroadcastStatusToPeersNode
-# ---------------------------------------------------------------------------
-
-
-class TestBroadcastStatusToPeersNode:
-    def test_returns_failure_without_trigger_activity(
-        self, populated_dl_with_peer
-    ):
-        """No trigger_activity_factory → FAILURE (BT-14-001).
-
-        Case has a peer recipient, so FilterPeerRecipientsNode returns
-        SUCCESS with a non-empty list.  CreateBroadcastActivityNode then
-        fails because no factory is available.
-        """
-        bridge = BTBridge(datalayer=populated_dl_with_peer)
-        node = BroadcastStatusToPeersNode(
-            status_id=STATUS_ID,
-            participant_id=PARTICIPANT_ID,
-            sender_actor_id=ACTOR_ID,
-            case_id=CASE_ID,
-        )
-        result = bridge.execute_with_setup(tree=node, actor_id=CASE_MANAGER_ID)
-        assert result.status == Status.FAILURE
-
-    def test_returns_success_when_no_peer_recipients(self, populated_bridge):
-        """No peer recipients (only vendor + Case Manager) → SUCCESS (no-op).
-
-        When the recipient list is empty after filtering, broadcast is not
-        needed and the whole subtree returns SUCCESS so downstream steps
-        (PublicDisclosureBranchNode, AutoCloseBranchNode) can still run.
-        """
-        node = BroadcastStatusToPeersNode(
-            status_id=STATUS_ID,
-            participant_id=PARTICIPANT_ID,
-            sender_actor_id=ACTOR_ID,
-            case_id=CASE_ID,
-        )
-        result = populated_bridge.execute_with_setup(
-            tree=node, actor_id=CASE_MANAGER_ID
-        )
-        assert result.status == Status.SUCCESS
-
-    def test_returns_failure_without_datalayer_data(self, bridge):
-        """case_id=None → FindCaseManagerNode FAILURE propagates (BT-14-001)."""
-        node = BroadcastStatusToPeersNode(
-            status_id=STATUS_ID,
-            participant_id=PARTICIPANT_ID,
-            sender_actor_id=ACTOR_ID,
-            case_id=None,
-        )
-        result = bridge.execute_with_setup(tree=node, actor_id=ACTOR_ID)
-        assert result.status == Status.FAILURE
-
-    def test_skips_when_current_actor_is_the_only_peer(self, populated_dl):
-        trigger_activity = MagicMock()
-        peer_actor_id = "https://example.org/actors/finder"
-        peer_participant_id = (
-            "https://example.org/cases/case-01/participants/finder"
-        )
-        peer_participant = CaseParticipant(
-            id_=peer_participant_id,
-            context=CASE_ID,
-            attributed_to=peer_actor_id,
-            case_roles=[CVDRole.FINDER],
-        )
-        populated_dl.create(peer_participant)
-        case = populated_dl.read(CASE_ID)
-        assert case is not None
-        case.actor_participant_index[peer_actor_id] = peer_participant_id
-        populated_dl.save(case)
-        bridge = BTBridge(
-            datalayer=populated_dl, trigger_activity=trigger_activity
-        )
-        node = BroadcastStatusToPeersNode(
-            status_id=STATUS_ID,
-            participant_id=PARTICIPANT_ID,
-            sender_actor_id=ACTOR_ID,
-            case_id=CASE_ID,
-        )
-        result = bridge.execute_with_setup(tree=node, actor_id=peer_actor_id)
-        assert result.status == Status.SUCCESS
-        trigger_activity.add_participant_status_to_participant.assert_not_called()
-
-    def test_broadcasts_from_case_manager_only(self, populated_dl):
-        trigger_activity = MagicMock()
-        trigger_activity.add_participant_status_to_participant.return_value = (
-            "urn:uuid:activity-1"
-        )
-        peer_actor_id = "https://example.org/actors/finder"
-        peer_participant_id = (
-            "https://example.org/cases/case-01/participants/finder"
-        )
-        peer_participant = CaseParticipant(
-            id_=peer_participant_id,
-            context=CASE_ID,
-            attributed_to=peer_actor_id,
-            case_roles=[CVDRole.FINDER],
-        )
-        populated_dl.create(peer_participant)
-        case = populated_dl.read(CASE_ID)
-        assert case is not None
-        case.actor_participant_index[peer_actor_id] = peer_participant_id
-        populated_dl.save(case)
-        bridge = BTBridge(
-            datalayer=populated_dl, trigger_activity=trigger_activity
-        )
-        node = BroadcastStatusToPeersNode(
-            status_id=STATUS_ID,
-            participant_id=PARTICIPANT_ID,
-            sender_actor_id=ACTOR_ID,
-            case_id=CASE_ID,
-        )
-        result = bridge.execute_with_setup(tree=node, actor_id=CASE_MANAGER_ID)
-        assert result.status == Status.SUCCESS
-        trigger_activity.add_participant_status_to_participant.assert_called_once_with(
-            status_id=STATUS_ID,
-            participant_id=PARTICIPANT_ID,
-            actor=CASE_MANAGER_ID,
-            to=[peer_actor_id],
-        )
-
-
-# ---------------------------------------------------------------------------
-# Step 3 leaf nodes: FindCaseManagerNode
-# ---------------------------------------------------------------------------
-
-
-PEER_ACTOR_ID = "https://example.org/actors/finder"
-PEER_PARTICIPANT_ID = "https://example.org/cases/case-01/participants/finder"
-
-
-@pytest.fixture
-def peer_participant():
-    return CaseParticipant(
-        id_=PEER_PARTICIPANT_ID,
-        context=CASE_ID,
-        attributed_to=PEER_ACTOR_ID,
-        case_roles=[CVDRole.FINDER],
-    )
-
-
-@pytest.fixture
-def populated_dl_with_peer(populated_dl, peer_participant):
-    """DataLayer with vendor, Case Manager, and a third peer participant."""
-    populated_dl.create(peer_participant)
-    case = populated_dl.read(CASE_ID)
-    assert case is not None
-    case.actor_participant_index[PEER_ACTOR_ID] = PEER_PARTICIPANT_ID
-    populated_dl.save(case)
-    return populated_dl
-
-
-class TestFindCaseManagerNode:
-    def test_returns_failure_without_case_id(self, populated_bridge):
-        """case_id=None → FAILURE (prerequisite missing)."""
-        node = FindCaseManagerNode(case_id=None)
-        result = populated_bridge.execute_with_setup(
-            tree=node, actor_id=CASE_MANAGER_ID
-        )
-        assert result.status == Status.FAILURE
-
-    def test_returns_failure_when_case_not_found(self, bridge):
-        """case_id set but no matching case in DataLayer → FAILURE."""
-        node = FindCaseManagerNode(case_id=CASE_ID)
-        result = bridge.execute_with_setup(tree=node, actor_id=CASE_MANAGER_ID)
-        assert result.status == Status.FAILURE
-
-    def test_returns_failure_when_no_case_manager(self, dl):
-        """Case exists but has no CASE_MANAGER participant → FAILURE."""
-        case_no_mgr = VulnerabilityCase(id_=CASE_ID, name="No Manager")
-        participant_no_mgr = CaseParticipant(
-            id_=PARTICIPANT_ID,
-            context=CASE_ID,
-            attributed_to=ACTOR_ID,
-            case_roles=[CVDRole.CASE_OWNER],
-        )
-        case_no_mgr.actor_participant_index[ACTOR_ID] = PARTICIPANT_ID
-        dl.create(case_no_mgr)
-        dl.create(participant_no_mgr)
-        bridge_no_mgr = BTBridge(datalayer=dl)
-        node = FindCaseManagerNode(case_id=CASE_ID)
-        result = bridge_no_mgr.execute_with_setup(tree=node, actor_id=ACTOR_ID)
-        assert result.status == Status.FAILURE
-
-    def test_returns_success_and_writes_blackboard(self, populated_bridge):
-        """Happy path: writes broadcast_case_manager_id to blackboard."""
-        node = FindCaseManagerNode(case_id=CASE_ID)
-        result = populated_bridge.execute_with_setup(
-            tree=node, actor_id=CASE_MANAGER_ID
-        )
-        assert result.status == Status.SUCCESS
-        assert (
-            py_trees.blackboard.Blackboard.storage[
-                "/broadcast_case_manager_id"
-            ]
-            == CASE_MANAGER_ID
-        )
-
-
-# ---------------------------------------------------------------------------
-# Step 3 leaf nodes: FilterPeerRecipientsNode
-# ---------------------------------------------------------------------------
-
-
-class TestFilterPeerRecipientsNode:
-    def _run_find_then_filter(self, dl, actor_id, sender_actor_id=ACTOR_ID):
-        """Helper: run FindCaseManagerNode then FilterPeerRecipientsNode."""
-        seq = py_trees.composites.Sequence(
-            name="TestSeq",
-            memory=False,
-            children=[
-                FindCaseManagerNode(case_id=CASE_ID),
-                FilterPeerRecipientsNode(
-                    sender_actor_id=sender_actor_id, case_id=CASE_ID
-                ),
-            ],
-        )
-        bridge = BTBridge(datalayer=dl)
-        return bridge.execute_with_setup(tree=seq, actor_id=actor_id)
-
-    def test_returns_failure_without_case_id(self, populated_dl):
-        """case_id=None → FAILURE from FilterPeerRecipientsNode."""
-        seq = py_trees.composites.Sequence(
-            name="TestSeq",
-            memory=False,
-            children=[
-                FindCaseManagerNode(case_id=CASE_ID),
-                FilterPeerRecipientsNode(
-                    sender_actor_id=ACTOR_ID, case_id=None
-                ),
-            ],
-        )
-        bridge = BTBridge(datalayer=populated_dl)
-        result = bridge.execute_with_setup(tree=seq, actor_id=CASE_MANAGER_ID)
-        assert result.status == Status.FAILURE
-
-    def test_returns_success_with_empty_list_when_no_eligible_recipients(
-        self, populated_dl
-    ):
-        """Only vendor and Case Manager in case → empty list, SUCCESS (no-op)."""
-        result = self._run_find_then_filter(
-            populated_dl,
-            actor_id=CASE_MANAGER_ID,
-            sender_actor_id=ACTOR_ID,
-        )
-        assert result.status == Status.SUCCESS
-        assert (
-            py_trees.blackboard.Blackboard.storage[
-                "/broadcast_peer_recipient_ids"
-            ]
-            == []
-        )
-
-    def test_returns_success_with_eligible_peer(self, populated_dl_with_peer):
-        """Third peer exists and is not sender/self/manager → SUCCESS."""
-        result = self._run_find_then_filter(
-            populated_dl_with_peer,
-            actor_id=CASE_MANAGER_ID,
-            sender_actor_id=ACTOR_ID,
-        )
-        assert result.status == Status.SUCCESS
-        assert py_trees.blackboard.Blackboard.storage[
-            "/broadcast_peer_recipient_ids"
-        ] == [PEER_ACTOR_ID]
-
-    def test_excludes_sender_from_recipients(self, populated_dl_with_peer):
-        """Peer is the sender → excluded; vendor (ACTOR_ID) is still eligible."""
-        result = self._run_find_then_filter(
-            populated_dl_with_peer,
-            actor_id=CASE_MANAGER_ID,
-            sender_actor_id=PEER_ACTOR_ID,
-        )
-        # ACTOR_ID is not the sender, not self (CASE_MANAGER_ID), and not the
-        # Case Manager, so it remains as an eligible recipient.
-        assert result.status == Status.SUCCESS
-        assert py_trees.blackboard.Blackboard.storage[
-            "/broadcast_peer_recipient_ids"
-        ] == [ACTOR_ID]
-
-    def test_excludes_self_from_recipients_empty_list(
-        self, populated_dl_with_peer
-    ):
-        """Actor running as the peer → peer excluded; empty list → SUCCESS (no-op)."""
-        result = self._run_find_then_filter(
-            populated_dl_with_peer,
-            actor_id=PEER_ACTOR_ID,
-            sender_actor_id=ACTOR_ID,
-        )
-        assert result.status == Status.SUCCESS
-        assert (
-            py_trees.blackboard.Blackboard.storage[
-                "/broadcast_peer_recipient_ids"
-            ]
-            == []
-        )
-
-
-# ---------------------------------------------------------------------------
-# Step 3 leaf nodes: BroadcastQueueToOutboxNode
-# ---------------------------------------------------------------------------
-
-
-class TestBroadcastQueueToOutboxNode:
-    def _run_full_broadcast_sequence(
-        self, dl, actor_id, trigger_activity=None
-    ):
-        """Run the full four-node broadcast inner sequence."""
-        seq = py_trees.composites.Sequence(
-            name="TestSeq",
-            memory=False,
-            children=[
-                FindCaseManagerNode(case_id=CASE_ID),
-                FilterPeerRecipientsNode(
-                    sender_actor_id=ACTOR_ID, case_id=CASE_ID
-                ),
-                CreateBroadcastActivityNode(
-                    activity_builder=lambda factory, cm_id, recipients: factory.add_participant_status_to_participant(
-                        status_id=STATUS_ID,
-                        participant_id=PARTICIPANT_ID,
-                        actor=cm_id,
-                        to=recipients,
-                    )
-                ),
-                BroadcastQueueToOutboxNode(),
-            ],
-        )
-        bridge = BTBridge(datalayer=dl, trigger_activity=trigger_activity)
-        return bridge.execute_with_setup(tree=seq, actor_id=actor_id)
-
-    def test_full_sequence_succeeds(self, populated_dl_with_peer):
-        """Full four-node broadcast sequence succeeds end-to-end."""
-        trigger_activity = MagicMock()
-        trigger_activity.add_participant_status_to_participant.return_value = (
-            "urn:uuid:activity-1"
-        )
-        result = self._run_full_broadcast_sequence(
-            populated_dl_with_peer,
-            actor_id=CASE_MANAGER_ID,
-            trigger_activity=trigger_activity,
-        )
-        assert result.status == Status.SUCCESS
-
-    def test_returns_failure_without_blackboard_keys(self):
-        """BroadcastQueueToOutboxNode returns FAILURE when broadcast blackboard
-        keys are missing (no prior FindCaseManagerNode run)."""
-        node = BroadcastQueueToOutboxNode()
-        dl_empty = SqliteDataLayer("sqlite:///:memory:")
-        bridge = BTBridge(datalayer=dl_empty)
-        # Without pre-populated blackboard keys the node returns FAILURE
-        result = bridge.execute_with_setup(tree=node, actor_id=CASE_MANAGER_ID)
-        assert result.status == Status.FAILURE
-
-
-# ---------------------------------------------------------------------------
 # Step 4: PublicDisclosureBranchNode
 # ---------------------------------------------------------------------------
 
@@ -1095,55 +732,6 @@ class TestPublicDisclosureBranchNode:
         )
 
         # State was still applied before the broadcast attempt
-        updated = c(VulnerabilityCase, populated_dl.read(CASE_ID))
-        assert updated.current_status.em_state == EM.EXITED
-        assert updated.active_embargo is None
-
-    def test_triggers_teardown_succeeds_with_factory(
-        self, populated_dl, case, status_obj
-    ):
-        """CS.P + CASE_OWNER sender + factory → SUCCESS and EM=EXITED."""
-        from unittest.mock import MagicMock
-        from typing import cast as c
-
-        from vultron.core.states.cs import CS_pxa
-        from vultron.core.states.em import EM
-        from vultron.wire.as2.vocab.objects.case_status import CaseStatus
-        from vultron.wire.as2.vocab.objects.embargo_event import EmbargoEvent
-        from vultron.wire.as2.vocab.objects.vulnerability_case import (
-            VulnerabilityCase,
-        )
-
-        embargo = EmbargoEvent(
-            id_=f"{CASE_ID}/embargo_events/e1", context=CASE_ID
-        )
-        case.active_embargo = embargo.id_
-        case.current_status.em_state = EM.ACTIVE
-        populated_dl.create(embargo)
-        populated_dl.save(case)
-
-        cs = CaseStatus()
-        cs.pxa_state = CS_pxa.Pxa
-        status_obj.case_status = cs
-        populated_dl.save(status_obj)
-
-        mock_factory = MagicMock()
-        mock_factory.terminate_embargo.return_value = (
-            "urn:uuid:terminate-1",
-            "urn:uuid:terminate-event-1",
-        )
-        bridge = BTBridge(
-            datalayer=populated_dl, trigger_activity=mock_factory
-        )
-
-        node = PublicDisclosureBranchNode(
-            status_obj=status_obj,
-            sender_actor_id=ACTOR_ID,
-            case_id=CASE_ID,
-        )
-        result = bridge.execute_with_setup(tree=node, actor_id=CASE_MANAGER_ID)
-        assert result.status == Status.SUCCESS
-
         updated = c(VulnerabilityCase, populated_dl.read(CASE_ID))
         assert updated.current_status.em_state == EM.EXITED
         assert updated.active_embargo is None
