+++ vultron/core/behaviors/status/nodes/broadcast.py
@@ -13,88 +13,38 @@
 #  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
 #  U.S. Patent and Trademark Office by Carnegie Mellon University
 
-"""Status-domain peer broadcast nodes for the AddParticipantStatus workflow.
+"""Case-manager lookup helper used by the status lifecycle nodes.
 
-Re-exports the shared generic nodes from
-:mod:`vultron.core.behaviors.broadcast.nodes` and provides the
-status-specific :class:`BroadcastStatusToPeersNode`.
+The raw peer re-broadcast step (step 3 of DEMOMA-07-003) was removed in
+DEMOMA-07-005; ``Announce(CaseLedgerEntry)`` fan-out in
+``_commit_log_cascade_bt`` already propagates status updates to all
+participants.
 
-Per DEMOMA-07-003 step 3 and specs/behavior-tree-integration.yaml
-BT-14-001, BT-14-002.
+:func:`_find_case_manager_id` remains here because
+:mod:`vultron.core.behaviors.status.nodes.lifecycle` uses it to resolve
+the Case Manager actor ID for auto-close fan-out.
 """
 
 import logging
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
-import py_trees
+from vultron.core.states.roles import CVDRole
 
-from vultron.core.behaviors.broadcast.nodes import (  # noqa: F401
-    BroadcastQueueToOutboxNode,
-    CreateBroadcastActivityNode,
-    FilterPeerRecipientsNode,
-    FindCaseManagerNode,
-    _find_case_manager_id,
-)
+if TYPE_CHECKING:
+    from vultron.core.ports.case_persistence import CasePersistence
 
 logger = logging.getLogger(__name__)
 
 
-class BroadcastStatusToPeersNode(py_trees.composites.Sequence):
-    """Step 3: Fail-fast broadcast of Add(ParticipantStatus) to peers.
-
-    The current actor re-sends the status update to all other participants,
-    excluding the original sender, itself, and the Case Manager.
-
-    Implemented as a ``py_trees.composites.Sequence`` (memory=False) via
-    :func:`~vultron.core.behaviors.broadcast.peer_broadcast_tree.peer_broadcast_bt`.
-    Returns FAILURE when broadcast preparation or outbox enqueueing fails
-    (BT-14-001).  Returns SUCCESS when the recipient list is empty (no peers
-    to notify — broadcast is a no-op, not a failure).
-
-    Broadcast sequence:
-
-    .. code-block:: text
-
-        ├── FindCaseManagerNode          # resolve Case Manager → blackboard
-        ├── FilterPeerRecipientsNode     # exclude sender/self/manager → blackboard
-        ├── CreateBroadcastActivityNode  # factory call → blackboard
-        └── BroadcastQueueToOutboxNode   # queue activity to Case Manager outbox
-
-    Per DEMOMA-07-003 step 3 and specs/behavior-tree-integration.yaml
-    BT-14-001, BT-14-002.
-    """
-
-    def __init__(
-        self,
-        status_id: str,
-        participant_id: str,
-        sender_actor_id: str,
-        case_id: str | None,
-        name: str | None = None,
-    ):
-        super().__init__(name=name or self.__class__.__name__, memory=False)
-
-        def _status_activity_builder(
-            factory: Any, case_manager_id: str, recipient_ids: list[str]
-        ) -> str:
-            result: str = factory.add_participant_status_to_participant(
-                status_id=status_id,
-                participant_id=participant_id,
-                actor=case_manager_id,
-                to=recipient_ids,
-            )
-            return result
-
-        self.add_children(
-            [
-                FindCaseManagerNode(case_id=case_id),
-                FilterPeerRecipientsNode(
-                    sender_actor_id=sender_actor_id,
-                    case_id=case_id,
-                ),
-                CreateBroadcastActivityNode(
-                    activity_builder=_status_activity_builder
-                ),
-                BroadcastQueueToOutboxNode(),
-            ]
-        )
+def _find_case_manager_id(dl: "CasePersistence", case: Any) -> str | None:
+    """Return the attributed_to actor ID for the CASE_MANAGER participant."""
+    for p_id in case.actor_participant_index.values():
+        p = dl.read(p_id)
+        if p is None:
+            continue
+        roles = getattr(p, "case_roles", [])
+        if CVDRole.CASE_MANAGER in roles:
+            attr = getattr(p, "attributed_to", None)
+            if attr:
+                return str(attr)
+    return None
